% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_jaccard}
\alias{calculate_jaccard}
\title{Calculate Jaccard Index}
\usage{
calculate_jaccard(set1, set2)
}
\arguments{
\item{set1}{Character vector}

\item{set2}{Character vector}
}
\value{
Numeric Jaccard index (0-1)
}
\description{
Computes Jaccard similarity between two sets.
}
