% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{counts_to_tpm}
\alias{counts_to_tpm}
\title{Convert Raw Counts to TPM}
\usage{
counts_to_tpm(counts, gene_length_info)
}
\arguments{
\item{counts}{Gene x Sample count matrix}

\item{gene_length_info}{Data frame with SYMBOL and length columns}
}
\value{
TPM normalized expression matrix
}
\description{
Converts raw count data to Transcripts Per Million (TPM).
}
\keyword{internal}
