% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_ALL_INPUT_GENES}
\alias{DEFAULT_ALL_INPUT_GENES}
\title{All Default LAMBclassifier Input Genes}
\format{
An object of class \code{character} of length 68.
}
\usage{
DEFAULT_ALL_INPUT_GENES
}
\description{
Consolidated list of all genes across default signatures.
}
\keyword{datasets}
