% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_APOCRINE_GENES}
\alias{DEFAULT_APOCRINE_GENES}
\title{Default Gene Signature: Apocrine}
\format{
An object of class \code{character} of length 14.
}
\usage{
DEFAULT_APOCRINE_GENES
}
\description{
Molecular apocrine carcinoma signature.
}
\keyword{datasets}
