% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_BASAL_LIKE_GENES}
\alias{DEFAULT_BASAL_LIKE_GENES}
\title{Default Gene Signature: Basal-like}
\format{
An object of class \code{character} of length 14.
}
\usage{
DEFAULT_BASAL_LIKE_GENES
}
\description{
Basal-like breast cancer signature.
}
\keyword{datasets}
