% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_CHONDROID_GENES}
\alias{DEFAULT_CHONDROID_GENES}
\title{Default Gene Signature: Chondroid}
\format{
An object of class \code{character} of length 9.
}
\usage{
DEFAULT_CHONDROID_GENES
}
\description{
Chondroid metaplastic carcinoma signature.
}
\keyword{datasets}
