% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{default_gene_sets}
\alias{default_gene_sets}
\title{Default LAMB Gene Sets}
\usage{
default_gene_sets()
}
\value{
Named list of character vectors.
}
\description{
The six signatures used by \code{\link{LAMBclassifier}}.
}
