% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_LHS_GENES}
\alias{DEFAULT_LHS_GENES}
\title{Default Gene Signature: Luminal Hormone Sensing (LHS)}
\format{
An object of class \code{character} of length 15.
}
\usage{
DEFAULT_LHS_GENES
}
\description{
Luminal hormone sensing (LHS) signature.
}
\keyword{datasets}
