% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_MALO_HER2_GENES}
\alias{DEFAULT_MALO_HER2_GENES}
\title{Default Gene Signature: MA-Lo (HER2)}
\format{
An object of class \code{character} of length 9.
}
\usage{
DEFAULT_MALO_HER2_GENES
}
\description{
ERBB2-amplicon expression and co-target signature (including
FGFR4 and PRODH). This is not a direct DNA-amplification assay.
}
\keyword{datasets}
