% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_METAPLASTIC_GENES}
\alias{DEFAULT_METAPLASTIC_GENES}
\title{Default Gene Signature: Metaplastic (Union)}
\format{
An object of class \code{character} of length 18.
}
\usage{
DEFAULT_METAPLASTIC_GENES
}
\description{
Union of DEFAULT_CHONDROID_GENES and DEFAULT_SQUAMOUS_EPITHELIAL_GENES.
}
\keyword{datasets}
