% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_SQUAMOUS_EPITHELIAL_GENES}
\alias{DEFAULT_SQUAMOUS_EPITHELIAL_GENES}
\title{Default Gene Signature: Squamous Epithelial}
\format{
An object of class \code{character} of length 9.
}
\usage{
DEFAULT_SQUAMOUS_EPITHELIAL_GENES
}
\description{
Squamous metaplastic carcinoma signature.
}
\keyword{datasets}
