% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAMBclassifier.R
\name{get_gene_length}
\alias{get_gene_length}
\title{Get gene length information from internal data}
\usage{
get_gene_length()
}
\description{
Get gene length information from internal data
}
\keyword{internal}
