% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{lamb_calibration}
\alias{lamb_calibration}
\title{Calibration Constants}
\usage{
lamb_calibration()
}
\value{
A list of calibration constants.
}
\description{
Decision thresholds used by \code{\link{LAMBclassifier}}.
}
