% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{LAMB_CLASS_COLORS}
\alias{LAMB_CLASS_COLORS}
\alias{LAB_CLASS_COLORS}
\title{LAMBclassifier Class Colours}
\format{
An object of class \code{character} of length 7.

An object of class \code{character} of length 7.
}
\usage{
LAMB_CLASS_COLORS

LAB_CLASS_COLORS
}
\description{
Colour palette for the LAMB classes.
}
\keyword{datasets}
