% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{LAMB_CLASS_ORDER}
\alias{LAMB_CLASS_ORDER}
\alias{LAB_CLASS_ORDER}
\title{LAMBclassifier Class Order}
\format{
An object of class \code{character} of length 7.

An object of class \code{character} of length 7.
}
\usage{
LAMB_CLASS_ORDER

LAB_CLASS_ORDER
}
\description{
Display order for the six LAMB classes.
}
\keyword{datasets}
