% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{lamb_universe}
\alias{lamb_universe}
\title{Reference Gene Universe}
\usage{
lamb_universe()
}
\value{
Character vector of HGNC symbols.
}
\description{
The fixed gene space in which all signature scores are computed.
}
