% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{LAMBclassifier-package}
\alias{LAMBclassifier-package}
\title{LAMBclassifier: Luminal-Apocrine-Metaplastic-Basal Breast Cancer Classifier}
\description{
A breast cancer sample classifier based on lineage-aware mammary breast
classification. Evaluates single-sample rank enrichment across six developmental
endpoints.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/debjeetrayx/LAMBclassifier}
  \item Report bugs at \url{https://github.com/debjeetrayx/LAMBclassifier/issues}
}

}
\author{
\strong{Maintainer}: Debjeet Ray \email{debjeet.ray@etu.u-bordeaux.fr}

}
\keyword{internal}
