\name{pam50.robust}
\alias{pam50.robust}
\docType{data}
\title{PAM50 Centroids Data}
\description{
PAM50 molecular subtype centroids for breast cancer classification.
Used internally by LAMBclassifier for PAM50 subtyping.
}
\usage{data(pam50.robust)}
\format{
A list containing PAM50 centroid values and gene mappings.
}
\source{
genefu package
}
\keyword{datasets}
