% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{PAM50_COLORS}
\alias{PAM50_COLORS}
\title{PAM50 Class Colours}
\format{
An object of class \code{character} of length 5.
}
\usage{
PAM50_COLORS
}
\description{
Colour palette for PAM50 intrinsic subtypes.
}
\keyword{datasets}
