% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{PAM50_ORDER}
\alias{PAM50_ORDER}
\title{PAM50 Class Order}
\format{
An object of class \code{character} of length 5.
}
\usage{
PAM50_ORDER
}
\description{
Standard display order for PAM50 intrinsic subtypes.
}
\keyword{datasets}
