% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_expression_scatter}
\alias{plot_expression_scatter}
\title{Plot Expression Scatter}
\usage{
plot_expression_scatter(norm_data, results, g1, g2, output_prefix)
}
\arguments{
\item{norm_data}{Expression matrix.}

\item{results}{Data frame from LAMBclassifier().}

\item{g1}{X-axis gene symbol.}

\item{g2}{Y-axis gene symbol.}

\item{output_prefix}{File path prefix.}
}
\description{
Creates a 2D scatter plot of two genes, colored by LAMB class.
}
