% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAMBclassifier.R
\name{run_pam50_internal}
\alias{run_pam50_internal}
\title{Run PAM50 Intrinsic Subtyping (Internal)}
\usage{
run_pam50_internal(data, pam50_centroids, gene_length_info)
}
\description{
Run PAM50 Intrinsic Subtyping (Internal)
}
\keyword{internal}
