% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_lamb_results}
\alias{summary_lamb_results}
\title{Summary of LAMBclassifier Results}
\usage{
summary_lamb_results(results)
}
\arguments{
\item{results}{Data frame returned by LAMBclassifier()}
}
\value{
Invisible NULL, prints summary to console
}
\description{
Prints a formatted summary of classification results.
}
\examples{
\dontrun{
results <- LAMBclassifier(data)
summary_lamb_results(results)
}

}
